Refactor proof_completion into a self-contained expert package; drop config
Consolidate the expert's code into one package under modules/ and remove the
experts.json config entirely — registration is now purely decorator-driven.

- modules/proof_completion/ now holds module, signature, outputs, handler,
  metric, models, graph_ops, grounding, dataset, domains/ (was split across
  modules/, handlers/, signatures.py, outputs.py, metrics.py, proof_completion/).
- Top level keeps only generic framework: registry, context_id, service,
  llm_config, outputs (Output base).
- discover() imports expert *packages* (modules/<pkg>); each package __init__
  imports its submodules so @register_* fire. No load_config/validate.
- Removed experts.json, metrics.py, signatures.py, handlers/.
- Updated scripts + tests to the new import paths; README rewritten.

46 tests pass; generator works across all 6 domains.
