--- backend/experts/handlers/__init__.py
@@ -1,1 +1,0 @@
-"""Output handlers. Each file self-registers via ``@register_handler``."""
