+++ backend/experts/modules/proof_completion/dataset.py
@@ -22,8 +22,8 @@ import sympy as sp
 from backend.model.semantic_graph import SemanticGraph
 from backend.semantic_graph.service import SemanticGraphService
 
-from ..context_id import build as build_context_id
-from ..outputs import GRAPH_OP_ADAPTER
+from backend.experts.context_id import build as build_context_id
+from .outputs import GRAPH_OP_ADAPTER
 from .graph_ops import apply, canonical_equal, diff
 from .grounding import is_grounded
 from .models import GraphTransition
