--- backend/experts/proof_completion/__init__.py
@@ -1,1 +1,0 @@
-"""ProofCompletionExpert: graph-transition models, op engine, metric, dataset."""
